+++ fixtures/gists-app/app/cookies.ts
@@ -1,0 +1,6 @@
+import { createCookie } from "@remix-run/data";
+
+export let userPrefsCookie = createCookie("user-prefs", {
+  path: "/",
+  httpOnly: false
+});