+++ packages/react/index.tsx
@@ -4,11 +4,14 @@ export {
   Styles,
   Link,
   Form,
+  useFormAction,
+  useSubmit,
   usePendingFormSubmit,
   useRouteData,
   usePendingLocation,
   useBeforeUnload
 } from "./components";
-export type { FormProps } from "./components";
+
+export type { FormProps, SubmitFunction } from "./components";
 
 export type { FormMethod, FormEncType, FormSubmit } from "./data";